Make lock guard movable & accept generalized arguments
